requirements.txt
annotated-types==0.6.0
blinker==1.7.0
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
diskcache==5.6.3
filelock==3.13.1
Flask==3.0.2
Flask-Cors==4.0.0
fsspec==2024.2.0
huggingface-hub==0.20.3
idna==3.6
interegular==0.3.3
itsdangerous==2.1.2
Jinja2==3.1.3
llama_cpp_python==0.2.50
lm-format-enforcer==0.9.0
MarkupSafe==2.1.5
mpmath==1.3.0
networkx==3.2.1
numpy==1.26.4
packaging==23.2
pydantic==2.6.2
pydantic-to-typescript==1.0.10
pydantic_core==2.16.3
PyYAML==6.0.1
regex==2023.12.25
requests==2.31.0
safetensors==0.4.2
sympy==1.12
tokenizers==0.15.2
torch==2.2.1
tqdm==4.66.2
transformers==4.38.1
typing_extensions==4.9.0
untruncate-json==1.0.0
urllib3==2.2.1
Werkzeug==3.0.1
